server/requirements.txt
tinydb
inbox.py